Inline RPC response-check helpers into rpc module
