+++ crates/rust-client/src/rpc/mod.rs
@@ -43,7 +43,7 @@
 
 use alloc::boxed::Box;
 use alloc::collections::{BTreeMap, BTreeSet};
-use alloc::string::String;
+use alloc::string::{String, ToString};
 use alloc::vec::Vec;
 use core::fmt;
 
@@ -83,8 +83,6 @@ pub use domain::limits::RpcLimits;
 pub use domain::status::{NetworkNoteStatus, NetworkNoteStatusInfo, RpcStatusInfo};
 pub use endpoint::Endpoint;
 
-mod verification;
-
 #[cfg(not(feature = "testing"))]
 mod generated;
 #[cfg(feature = "testing")]
@@ -116,6 +114,89 @@ pub enum AccountStateAt {
 /// in any header slot indicates the note has attachments whose content must be fetched separately.
 fn metadata_has_attachments(metadata: &NoteMetadata) -> bool {
     metadata.attachment_headers().iter().any(|header| header.scheme().is_some())
+}
+
+// RESPONSE VERIFICATION
+// ================================================================================================
+//
+// Helpers used by the trait's default methods to check that a node's reply matches what was
+// requested, returning [`RpcError::InvalidResponse`] on a mismatch.
+
+/// Returns [`RpcError::InvalidResponse`] if `requested` is `Some` and `returned` does not equal it.
+fn verify_block_num(requested: Option<BlockNumber>, returned: BlockNumber) -> Result<(), RpcError> {
+    if let Some(requested) = requested
+        && returned != requested
+    {
+        return Err(RpcError::InvalidResponse(format!(
+            "node returned block {returned} but block {requested} was requested"
+        )));
+    }
+    Ok(())
+}
+
+/// Returns [`RpcError::InvalidResponse`] if any returned note ID was not in `requested`.
+fn verify_note_ids(
+    requested: &BTreeSet<NoteId>,
+    returned: impl IntoIterator<Item = NoteId>,
+) -> Result<(), RpcError> {
+    for id in returned {
+        if !requested.contains(&id) {
+            let list = requested.iter().map(ToString::to_string).collect::<Vec<_>>().join(", ");
+            return Err(RpcError::InvalidResponse(format!(
+                "node returned note {id} but [{list}] were requested"
+            )));
+        }
+    }
+    Ok(())
+}
+
+/// Returns [`RpcError::InvalidResponse`] if any returned note tag was not in `requested`.
+fn verify_note_tags(
+    requested: &BTreeSet<NoteTag>,
+    returned: impl IntoIterator<Item = NoteTag>,
+) -> Result<(), RpcError> {
+    for tag in returned {
+        if !requested.contains(&tag) {
+            let list = requested.iter().map(ToString::to_string).collect::<Vec<_>>().join(", ");
+            return Err(RpcError::InvalidResponse(format!(
+                "node returned note with tag {tag} but [{list}] were requested"
+            )));
+        }
+    }
+    Ok(())
+}
+
+/// Returns [`RpcError::InvalidResponse`] if any update carries a nullifier whose prefix was not in
+/// `requested_prefixes`.
+fn verify_nullifier_prefixes(
+    requested_prefixes: &BTreeSet<u16>,
+    batch: &[NullifierUpdate],
+) -> Result<(), RpcError> {
+    for update in batch {
+        let prefix = update.nullifier.prefix();
+        if !requested_prefixes.contains(&prefix) {
+            let requested = requested_prefixes
+                .iter()
+                .map(ToString::to_string)
+                .collect::<Vec<_>>()
+                .join(", ");
+            return Err(RpcError::InvalidResponse(format!(
+                "node returned nullifier with prefix {prefix} but [{requested}] were requested"
+            )));
+        }
+    }
+    Ok(())
+}
+
+/// Returns [`RpcError::InvalidResponse`] if `script`'s root does not equal the `requested` root.
+fn verify_note_script_root(requested: Word, script: &NoteScript) -> Result<(), RpcError> {
+    let fetched_root = script.root();
+    if Word::from(fetched_root) != requested {
+        return Err(RpcError::InvalidResponse(format!(
+            "node returned note script with root {fetched_root} for requested root {requested}"
+        )));
+    }
+    Ok(())
 }
 
 // NODE RPC CLIENT TRAIT
@@ -191,7 +272,7 @@ pub trait NodeRpcClient: Send + Sync {
     ) -> Result<(BlockHeader, Option<MmrProof>), RpcError> {
         let (header, mmr_proof) =
             self.get_block_header_by_number_unchecked(block_num, include_mmr_proof).await?;
-        verification::verify_block_num(block_num, header.block_num())?;
+        verify_block_num(block_num, header.block_num())?;
         Ok((header, mmr_proof))
     }
 
@@ -219,7 +300,7 @@ pub trait NodeRpcClient: Send + Sync {
         include_proof: bool,
     ) -> Result<ProvenBlock, RpcError> {
         let block = self.get_block_by_number_unchecked(block_num, include_proof).await?;
-        verification::verify_block_num(Some(block_num), block.header().block_num())?;
+        verify_block_num(Some(block_num), block.header().block_num())?;
         Ok(block)
     }
 
@@ -251,7 +332,7 @@ pub trait NodeRpcClient: Send + Sync {
     async fn get_notes_by_id(&self, note_ids: &[NoteId]) -> Result<Vec<FetchedNote>, RpcError> {
         let notes = self.get_notes_by_id_unchecked(note_ids).await?;
         let requested: BTreeSet<NoteId> = note_ids.iter().copied().collect();
-        verification::verify_note_ids(&requested, notes.iter().map(FetchedNote::id))?;
+        verify_note_ids(&requested, notes.iter().map(FetchedNote::id))?;
         Ok(notes)
     }
 
@@ -341,7 +422,7 @@ pub trait NodeRpcClient: Send + Sync {
         note_tags: &BTreeSet<NoteTag>,
     ) -> Result<Vec<NoteSyncBlock>, RpcError> {
         let blocks = self.sync_notes_unchecked(block_from, block_to, note_tags).await?;
-        verification::verify_note_tags(
+        verify_note_tags(
             note_tags,
             blocks.iter().flat_map(|b| b.notes.values().map(CommittedNote::tag)),
         )?;
@@ -428,7 +509,7 @@ pub trait NodeRpcClient: Send + Sync {
     ) -> Result<Vec<NullifierUpdate>, RpcError> {
         let nullifiers = self.sync_nullifiers_unchecked(prefix, block_from, block_to).await?;
         let requested: BTreeSet<u16> = prefix.iter().copied().collect();
-        verification::verify_nullifier_prefixes(&requested, &nullifiers)?;
+        verify_nullifier_prefixes(&requested, &nullifiers)?;
         Ok(nullifiers)
     }
 
@@ -472,7 +553,7 @@ pub trait NodeRpcClient: Send + Sync {
             AccountStateAt::ChainTip => None,
         };
         let (block_num, proof) = self.get_account_unchecked(account_id, request).await?;
-        verification::verify_block_num(requested, block_num)?;
+        verify_block_num(requested, block_num)?;
         Ok((block_num, proof))
     }
 
@@ -649,7 +730,7 @@ pub trait NodeRpcClient: Send + Sync {
     async fn get_note_script_by_root(&self, root: Word) -> Result<Option<NoteScript>, RpcError> {
         let script = self.get_note_script_by_root_unchecked(root).await?;
         if let Some(script) = &script {
-            verification::verify_note_script_root(root, script)?;
+            verify_note_script_root(root, script)?;
         }
         Ok(script)
     }
@@ -798,5 +879,78 @@ impl fmt::Display for RpcEndpoint {
             RpcEndpoint::GetLimits => write!(f, "get_limits"),
             RpcEndpoint::GetNetworkNoteStatus => write!(f, "get_network_note_status"),
         }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use core::slice;
+    use std::collections::BTreeSet;
+
+    use miden_protocol::note::{NoteId, NoteTag, Nullifier};
+    use miden_protocol::{Felt, Word};
+
+    use super::{verify_note_ids, verify_note_tags, verify_nullifier_prefixes};
+    use crate::rpc::RpcError;
+    use crate::rpc::domain::nullifier::NullifierUpdate;
+
+    fn nullifier_with_prefix(prefix: u16) -> Nullifier {
+        Nullifier::from_raw(Word::new([
+            Felt::ZERO,
+            Felt::ZERO,
+            Felt::ZERO,
+            Felt::new_unchecked(u64::from(prefix) << 48),
+        ]))
+    }
+
+    #[test]
+    fn verify_nullifier_prefixes_rejects_unrequested() {
+        let requested = NullifierUpdate {
+            nullifier: nullifier_with_prefix(0x1234),
+            block_num: 1u32.into(),
+        };
+        let unrequested = NullifierUpdate {
+            nullifier: nullifier_with_prefix(0xabcd),
+            block_num: 2u32.into(),
+        };
+
+        let requested_prefixes: BTreeSet<u16> = BTreeSet::from([0x1234]);
+
+        verify_nullifier_prefixes(&requested_prefixes, slice::from_ref(&requested))
+            .expect("requested prefix must be accepted");
+
+        let err = verify_nullifier_prefixes(&requested_prefixes, &[requested, unrequested])
+            .expect_err("unrequested prefix must be rejected");
+        assert!(matches!(err, RpcError::InvalidResponse(_)));
+    }
+
+    #[test]
+    fn verify_note_tags_rejects_unrequested() {
+        let requested = NoteTag::new(1);
+        let other = NoteTag::new(2);
+        let requested_set = BTreeSet::from([requested]);
+
+        verify_note_tags(&requested_set, [requested]).expect("requested tag must be accepted");
+
+        let err = verify_note_tags(&requested_set, [other])
+            .expect_err("unrequested tag must be rejected");
+        assert!(matches!(err, RpcError::InvalidResponse(_)));
+    }
+
+    fn note_id(n: u32) -> NoteId {
+        NoteId::from_raw(Word::from([n, 0, 0, 0]))
+    }
+
+    #[test]
+    fn verify_note_ids_rejects_unrequested() {
+        let requested = note_id(1);
+        let other = note_id(2);
+        let requested_set = BTreeSet::from([requested]);
+
+        verify_note_ids(&requested_set, [requested]).expect("requested note id must be accepted");
+
+        let err = verify_note_ids(&requested_set, [other])
+            .expect_err("unrequested note id must be rejected");
+        assert!(matches!(err, RpcError::InvalidResponse(_)));
     }
 }
